fix: error when trying to delegate from ledger (#83)
Clicking the delegate button from a validator page when having connected
a ledger device is reporting a generic error "something goes wrong..."

With debug mode active, it shows a `TypeError: Cannot read property
'unbonding_time' of null` from the validator page component.

Root issue comes from the sync service, which doesn't pull and store the
chain staking parameters from the RPC endpoint. Previously, it was taken
from the genesis file, but this mean it wouldn't be updated in case of
governance vote to change them.

This add a proper pull from RPC of the staking parameters, storing them
at the expected location for the Ledger action buttons to work.